# tests/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/payouts": {
      "post": {
        "summary": "Request Payout",
        "operationId": "request_payout_payouts_post",
        "parameters": [
          {
            "name": "X-Approval-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Approval-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/payouts/{run_id}/retry": {
      "post": {
        "summary": "Retry Payout",
        "description": "schema-contract.md §8 재발송 — FAILED run을 새 승인 토큰으로 다시 보낸다.\n사람이 먼저 POST /settlements/runs/{run_id}/approve로 FAILED→APPROVED 재승인해\n새 토큰을 받아야 한다(재발송도 돈이 나가는 행위라 approve의 CAS·캡 검사를 다시 탄다).\n이 라우트는 그 토큰으로 APPROVED→EXECUTING만 담당한다 — verify_and_burn_token은\n상태가 APPROVED이기만 하면 최초 송금과 재발송을 구분하지 않는다.",
        "operationId": "retry_payout_payouts__run_id__retry_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "X-Approval-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Approval-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/paypal": {
      "post": {
        "summary": "Paypal Webhook",
        "description": "schema-contract.md §8 — 보조 경로. 서명 검증 없음: 샌드박스 웹훅이 불안정해\n데모를 여기 걸지 않는다(주 경로는 /tasks/reconcile 폴링). 웹훅이 오면 같은 대조\n로직을 앞당겨 실행할 뿐, 안 와도 폴링으로 결말이 난다.",
        "operationId": "paypal_webhook_webhooks_paypal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/execute-payout": {
      "post": {
        "summary": "Task Execute Payout",
        "description": "Cloud Tasks가 부르는 실제 PayPal 호출 지점. schema-contract.md §3/§8, money-safety.md\n멱등성 — sender_batch_id/sender_item_id는 (run_id, recipient_id, retry_seq)에서\n결정론적으로 만든다.\n\nrecipient가 여럿이면 단일 PayPal 배치에 recipient당 item 하나씩 담아 한 번에 보낸다\n(sender_batch_id는 recipient와 무관하게 하나 — build_payout_ids 참조). 각 item의\n금액은 run 승인 시점에 고정된 fx_rates로 그 recipient 몫만 환산·합산한 값이다.",
        "operationId": "task_execute_payout_tasks_execute_payout_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/reconcile": {
      "post": {
        "summary": "Task Reconcile",
        "description": "schema-contract.md §8 — 주 경로. /payouts enqueue 시 RECONCILE_DELAY_SECONDS 뒤에\nCloud Tasks가 부른다. 미종결(pending_items가 있고 아직 최대 시도 전)이면 이 라우트가\n스스로를 다시 예약한다 — reconcile()이 최대 시도(PAYOUT_MAX_RECONCILE_ATTEMPTS)를\n넘기면 강제로 종결 처리하므로, 그 안에서는 계속 폴링해도 된다.",
        "operationId": "task_reconcile_tasks_reconcile_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}